requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
python-multipart==0.0.6
Pillow==10.1.0
pydantic==2.5.0
openai==1.58.1
python-dotenv==1.0.1
# openai==1.58.1 expects httpx that still supports the `proxies=` kwarg.
# (httpx 0.28 removed it in favor of `proxy=`), which can crash the worker at runtime.
httpx==0.27.2

